+++ crates/ruma-client-api/src/r0/session.rs
@@ -6,5 +6,4 @@ pub mod login_fallback;
 pub mod logout;
 pub mod logout_all;
 pub mod sso_login;
-#[cfg(feature = "unstable-pre-spec")]
 pub mod sso_login_with_provider;